add German signals Ra 10 and Ra 11/11a/11b
This commit also adds the Austrian "Wartesignal" if it is as sign and no
light signal because it is the same signal as used in Germany.
This commit also adds the Austrian "Verschubhalttafel" because it is the same
as used in Germany. The small icons cannot contain text due to their size.