render track refs when usage is given
The layout is similar to the signs used in stations in Germany by Deutsche
Bahn. This makes them much more prominent on higher zoom levels. On zoom levels
below 17 the track refs are not drawn anymore. Normal and narrow_gauge rails
are now using the same highlighting, the differences can be seen on the lines
themselves.