Add support for new tagging scheme for Lf6/7 signals
Problem: When having two Lf7 signals at the same position showing in opposite direction, there is a hack needed having two nodes nearby.

With the proposed new tagging scheme a Lf 7 signal automatically shows in "both directions", having a nominal and reverse speed (regarding to the OSM way).

Additionally, the renderer should only render the nominal speed limit, if a reverse is tagged too.

This results in crystal clear renderings on lines with 2 tracks. Every track renders the Lf 7 only on the right track.

A solution for lines with only one track is missing here!

It is not neccessary to render Lf 6 Kennziffer 16-20, as there are not defined in Germany.

The advantage of the proposed tagging scheme railway:signal:speed_limit:announcement is to stick in only one namespace, namely railway:signal:speed_limit.

There is no neccess to have a seperate namespace railway:signal:speed_limit_distant.