reorder German speed signals
This changes the z-index for Lf6/Lf7 and Zs3v/Zs3 signals so that the actual
signals (not the distant ones) have the higher index. While the distant signals
are probably more important for someone actually driving the train, for the map
the "real" signals are the more interesting ones. They need to be reordered in
the file, so that the ones with higher z-index come last so that multiple
signals at the same pole are rendered correctly.