fix generated permalink if the base URL contains subdirectories
If the base URL is not the server root the subdirectories were used from both
the root variable and the current URL, leading to duplication of the
directories which resulted in broken links.

Example:
  root = http://example.com/test/
  URL was http://example.com/test/index.php
  generated permalink was http://example.com/test/test/index.php

Simply use hostname and path from the current URL as suggested by rurseekatze to
avoid this.