JOSM presets: Removed defaults to avoid wrong data
Removed most of the default values in the JOSM presets to avoid wrong data.
Example: If you use the German infrastructure preset and do not change the selection boxes for railway:pzb=* or railway:lzb=*, it tags railway:pzb=no which can be wrong.
If a user does not enter anything, this information is unknown and not 'no'.