search: refactor milestone output
In case a milestone search is performed output all rows the same way. Until now
only signals, milestones, and level crossings were printed at all. There exist
objects in the database that fall into neither category but still have a
position on a track set, and which are also returned by the query.

Also remove the special case for level crossings, the type of each object is
printed later in the line anyway, which would lead to that text apprearing
twice.