maxspeed legend: do not show Zs10 light signal
As of today there is only one instance of this signals in the entire OSM
database. Several discussions on the net suggest that this may as well be a
mistagging, as the common narrative is that this signal type has been defined,
but never actually installed, and there will be no new installations as Zs3 will
be used instead. Keep the rendering for the moment, but don't waste space in the
already crowded legend.