Change title of the imprint page to "imprint"
The page contains important information according to §5 TMG (German
Telemedia Act). The link to it needs to be find and the link title
should be pretty clear that the imprint (German "Impressum") is behind
that link. There is pretty much case law about that.

Coypright topics are mentioned on that page as well because they more or
less fit there as another legal topic as well as privacy terms.