Add cones for signal direction and arrows for position
JOSM >= version 18678 supports parent_way_angle() to use the direction
of the OSM way to rotate node icons. This can be used to render cones
like iD and to arrows pointing to the left/right.

Thanks to @Tordanik for pointing me to that feature. See also
https://github.com/tordanik/direction-cones-style for a JOSM style
supporting direction=N/S/E/… and direction=<degrees>.